Fixed ONE bug. Therefore expecting 20 new bugs.

+++ Test Graphics with Physics/src/Physics/PhysicalFace.java
@@ -21,7 +21,9 @@ public class PhysicalFace {
 	public boolean collidesWithBall(Ball b) {
 		dist.set((point1.x - b.getPosition().x), (point1.y - b.getPosition().y), (point1.z - b.getPosition().z));
 		double distance = Math.abs(Vector3f.dot(normal, dist))/normal.length();
-		if ((float)distance <= b.getRadius() && bbox.inBoundingBox(b)) {
+		//System.out.println("Distance between ball and face: " + distance + " (radius of the ball = " + b.getRadius() + ")");
+		if (distance <= b.getRadius() && bbox.inBoundingBox(b)) {
+			//System.out.println("Collision between face and ball detected");
 			return true;
 		}
 		return false;
